+++ src/wasm_runtime/build.rs
@@ -60,6 +60,7 @@ fn main() {
     cfg.compile("wasm_runtime");
 
     println!("cargo::rerun-if-env-changed=WIT_WORLD");
+    println!("cargo::rerun-if-env-changed=WIT_WORLD_NAME");
     println!("cargo::rustc-check-cfg=cfg(component)");
     if env::var_os("WIT_WORLD").is_some() {
         println!("cargo::rustc-cfg=component");